+++ appinfo/application.php
@@ -44,7 +44,7 @@ class Application extends App {
 	 *
 	 * @param array $urlParams
 	 */
-	public function __construct(array $urlParams = array()) {
+	public function __construct(array $urlParams = []) {
 		parent::__construct('galleryplus', $urlParams);
 
 		$container = $this->getContainer();