+++ appinfo/routes.php
@@ -26,115 +26,115 @@ namespace OCA\GalleryPlus\AppInfo;
 $application = new Application();
 
 $application->registerRoutes(
-	$this, array(
-			 'routes' => array(
+	$this, [
+			 'routes' => [
 				 /**
 				  * Main routes
 				  */
-				 array(
+				 [
 					 'name' => 'page#index',
 					 'url'  => '/',
 					 'verb' => 'GET'
-				 ),
+				 ],
 				 // Landing page for public galleries
-				 array(
+				 [
 					 'name' => 'page#public_index',
 					 'url'  => '/s/{token}',
 					 'verb' => 'GET'
-				 ),
+				 ],
 				 // Landing page after password entry, for public galleries
-				 array(
+				 [
 					 'name' => 'page#public_index',
 					 'url'  => '/s/{token}',
 					 'verb' => 'POST',
 					 'postfix' => 'post'
-				 ),
+				 ],
 				 // An error page which can handle different error codes
-				 array(
+				 [
 					 'name' => 'page#error_page',
 					 'url'  => '/error',
 					 'verb' => 'GET'
-				 ),
+				 ],
 				 // The same page, but for POST
-				 array(
+				 [
 					 'name' => 'page#error_page',
 					 'url'  => '/error',
 					 'verb' => 'POST',
 					 'postfix' => 'post'
-				 ),
+				 ],
 				 /**
 				  * Services
 				  */
 				 // Album creation. Only called once logged in
-				 array(
+				 [
 					 'name' => 'service#get_album_info',
 					 'url'  => '/albums',
 					 'verb' => 'GET'
-				 ),
+				 ],
 				 // Supported media types. Only called by the slideshow
-				 array(
+				 [
 					 'name' => 'service#get_types',
 					 'url'  => '/mimetypes',
 					 'verb' => 'GET'
-				 ),
+				 ],
 				 // All the images of which a preview can be generated
-				 array(
+				 [
 					 'name' => 'service#get_images',
 					 'url'  => '/images',
 					 'verb' => 'GET'
-				 ),
+				 ],
 				 // Batch creation of thumbnails
-				 array(
+				 [
 					 'name' => 'service#get_thumbnails',
 					 'url'  => '/thumbnails',
 					 'verb' => 'GET'
-				 ),
+				 ],
 				 // Large preview of a file
-				 array(
+				 [
 					 'name' => 'service#show_preview',
 					 'url'  => '/preview',
 					 'verb' => 'GET'
-				 ),
+				 ],
 				 // Download the file
-				 array(
+				 [
 					 'name' => 'service#download_preview',
 					 'url'  => '/download',
 					 'verb' => 'GET'
-				 ),
+				 ],
 				 /**
 				  * Public services
 				  */
-				 array(
+				 [
 					 'name' => 'public_service#get_types',
 					 'url'  => '/mimetypes.public',
 					 'verb' => 'GET'
-				 ),
-				 array(
+				 ],
+				 [
 					 'name' => 'public_service#get_images',
 					 'url'  => '/images.public',
 					 'verb' => 'GET'
-				 ),
-				 array(
+				 ],
+				 [
 					 'name' => 'public_service#get_thumbnails',
 					 'url'  => '/thumbnails.public',
 					 'verb' => 'GET'
-				 ),
-				 array(
+				 ],
+				 [
 					 'name' => 'public_service#show_preview',
 					 'url'  => '/preview.public',
 					 'verb' => 'GET'
-				 ),
-				 array(
+				 ],
+				 [
 					 'name' => 'public_service#download_preview',
 					 'url'  => '/download.public',
 					 'verb' => 'GET'
-				 ),
+				 ],
 				 // API, for later
-				 /*array(
+				 /*[
 					 'name' => 'api#get_types',
 					 'url'  => '/api/1.0/types',
 					 'verb' => 'GET'
-				 ),*/
-			 )
-		 )
+				 ],*/
+			 ]
+		 ]
 );