+++ templates/public.php
@@ -7,19 +7,19 @@
  */
 script(
 	$_['appName'],
-	array(
+	[
 		'album',
 		'gallery',
 		'thumbnail'
-	)
+	]
 );
 style(
 	$_['appName'],
-	array(
+	[
 		'styles',
 		'mobile',
 		'public'
-	)
+	]
 );
 
 ?>