+++ utility/smarterlogger.php
@@ -56,7 +56,7 @@ class SmarterLogger implements ILogger {
 	 *
 	 * @return null
 	 */
-	public function emergency($message, array $context = array()) {
+	public function emergency($message, array $context = []) {
 		$this->log(\OCP\Util::FATAL, $message, $context);
 	}
 
@@ -68,7 +68,7 @@ class SmarterLogger implements ILogger {
 	 *
 	 * @return null
 	 */
-	public function alert($message, array $context = array()) {
+	public function alert($message, array $context = []) {
 		$this->log(\OCP\Util::ERROR, $message, $context);
 	}
 
@@ -80,7 +80,7 @@ class SmarterLogger implements ILogger {
 	 *
 	 * @return null
 	 */
-	public function critical($message, array $context = array()) {
+	public function critical($message, array $context = []) {
 		$this->log(\OCP\Util::ERROR, $message, $context);
 	}
 
@@ -93,7 +93,7 @@ class SmarterLogger implements ILogger {
 	 *
 	 * @return null
 	 */
-	public function error($message, array $context = array()) {
+	public function error($message, array $context = []) {
 		$this->log(\OCP\Util::ERROR, $message, $context);
 	}
 
@@ -105,7 +105,7 @@ class SmarterLogger implements ILogger {
 	 *
 	 * @return null
 	 */
-	public function warning($message, array $context = array()) {
+	public function warning($message, array $context = []) {
 		$this->log(\OCP\Util::WARN, $message, $context);
 	}
 
@@ -117,7 +117,7 @@ class SmarterLogger implements ILogger {
 	 *
 	 * @return null
 	 */
-	public function notice($message, array $context = array()) {
+	public function notice($message, array $context = []) {
 		$this->log(\OCP\Util::INFO, $message, $context);
 	}
 
@@ -129,7 +129,7 @@ class SmarterLogger implements ILogger {
 	 *
 	 * @return null
 	 */
-	public function info($message, array $context = array()) {
+	public function info($message, array $context = []) {
 		$this->log(\OCP\Util::INFO, $message, $context);
 	}
 
@@ -141,7 +141,7 @@ class SmarterLogger implements ILogger {
 	 *
 	 * @return null
 	 */
-	public function debug($message, array $context = array()) {
+	public function debug($message, array $context = []) {
 		$this->log(\OCP\Util::DEBUG, $message, $context);
 	}
 
@@ -155,7 +155,7 @@ class SmarterLogger implements ILogger {
 	 *
 	 * @return mixed
 	 */
-	public function log($level, $message, array $context = array()) {
+	public function log($level, $message, array $context = []) {
 		array_walk($context, [$this, 'contextNormalizer']);
 
 		if (!isset($context['app'])) {